edit CurrentUserDetails
